# Optimise away calls to cipher.NewCTR