This avoids the allocation of a 512-byte buffer on each packet, which leads to a 30% to 50% speedup in the SRTP benchmaks.

Edit (28/04/2022): see https://github.com/pion/srtp/pull/76#issuecomment-1112323998 for updated benchmarks.

```
name                 old time/op    new time/op     delta
EncryptRTP-4           4.15µs ± 0%     2.85µs ± 1%  -31.38%  (p=0.000 n=8+7)
EncryptRTPInPlace-4    4.05µs ± 2%     2.75µs ± 0%  -32.19%  (p=0.000 n=7+8)
DecryptRTP-4           3.81µs ± 0%     2.02µs ± 1%  -47.01%  (p=0.000 n=8+8)
Write-4                4.23µs ± 1%     2.93µs ± 0%  -30.79%  (p=0.000 n=8+8)
WriteRTP-4             4.15µs ± 0%     2.86µs ± 1%  -31.10%  (p=0.000 n=7+8)

name                 old speed      new speed       delta
EncryptRTP-4         27.0MB/s ± 0%   39.3MB/s ± 1%  +45.71%  (p=0.000 n=8+7)
EncryptRTPInPlace-4  27.7MB/s ± 2%   40.8MB/s ± 0%  +47.46%  (p=0.000 n=7+8)
DecryptRTP-4         7.34MB/s ± 0%  13.86MB/s ± 1%  +88.73%  (p=0.000 n=8+8)
Write-4              26.5MB/s ± 1%   38.2MB/s ± 0%  +44.47%  (p=0.000 n=8+8)
WriteRTP-4           27.0MB/s ± 0%   39.2MB/s ± 1%  +45.16%  (p=0.000 n=7+8)

name                 old alloc/op   new alloc/op    delta
EncryptRTP-4             788B ± 0%       212B ± 0%  -73.10%  (p=0.000 n=8+8)
EncryptRTPInPlace-4      660B ± 0%        84B ± 0%  -87.27%  (p=0.000 n=8+8)
DecryptRTP-4             692B ± 0%       116B ± 0%  -83.24%  (p=0.000 n=8+8)
Write-4                  788B ± 0%       212B ± 0%  -73.10%  (p=0.000 n=8+8)
WriteRTP-4               788B ± 0%       212B ± 0%  -73.10%  (p=0.000 n=8+8)

name                 old allocs/op  new allocs/op   delta
EncryptRTP-4             7.00 ± 0%       6.00 ± 0%  -14.29%  (p=0.000 n=8+8)
EncryptRTPInPlace-4      6.00 ± 0%       5.00 ± 0%  -16.67%  (p=0.000 n=8+8)
DecryptRTP-4             7.00 ± 0%       6.00 ± 0%  -14.29%  (p=0.000 n=8+8)
Write-4                  7.00 ± 0%       6.00 ± 0%  -14.29%  (p=0.000 n=8+8)
WriteRTP-4               7.00 ± 0%       6.00 ± 0%  -14.29%  (p=0.000 n=8+8)