@Sean-Der 2020-06-02
I vote for merging it, but I think we should keep it out of the default code path. I worry that a mistake in this code could really damage Pion's reputation. With DTLS we didn't have an option, but for this if we mess it up it could damage user trust for ever. People love to harp on `do not roll your own crypto` and even if the code isn't in the real crypto stuff they would use this as the example :/

What do you think of making this controllable via a new `ContextOption`? We can then put this in the `SettingEngine` in pion/webrtc.

I think we should open a bug upstream and see what they say. Hopefully we can land something like this upstream, and then put the new API behind a version check!