@at-wat 2020-06-02
> `do not roll your own crypto`

I really thought it.
At least, such code must be tested for various possible (especially erroneous) input patterns.
Making this controllable via a new `ContextOption` would be good for now, I think.