@jech 2020-06-02
> I vote for merging it, but I think we should keep it out of the default code path

I understand your hesitation, but I'm not sure I agree.

This is not "rolling your own crypto".  I'm just reimplementing a well-known, standard, boring piece of crypto machinery that has seen extensive peer review and that is very well understood.  This is completely different from designing new crypto algorithms.  (If we cannot even reimplement well-known algorithms, then why does Go have its own crypto stack rather than using OpenSSL?)

Second, any issue will be readily apparent: there is a test that it produces the same values for all reasonable packet sizes as the reference implementation.

Third, any security issues in this code will be readily apparent even without the above test — it's not anything subtle like checking MACs or protecting against replay, it's just stream encryption.  If something goes wrong notwithstanding the test above, then the symptoms will be incorrectly encrypted or decrypted packets which will be dropped.

> What do you think of making this controllable via a new `ContextOption`?

Oh, that's overkill.  If for some reason you want to make it user controllable, just define a global variable that can be initialised in `main`.

> I think we should open a bug upstream and see what they say.

 You do it or I do?