@jech 2022-04-28
I've revived this patch.  Updated benchmarks follow.  @Sean-Der please consider for merging, @davidzhao please provide encouragement ;-)

```
name                        old time/op    new time/op     delta
EncryptRTP-8                  1.58µs ± 8%     1.06µs ± 2%   -33.13%  (p=0.029 n=4+4)
EncryptRTPInPlace-8           1.54µs ± 3%     1.00µs ± 1%   -34.82%  (p=0.029 n=4+4)
DecryptRTP-8                  1.44µs ± 6%     0.67µs ± 2%   -53.80%  (p=0.029 n=4+4)
Write-8                       1.65µs ±10%     1.08µs ± 1%   -34.60%  (p=0.029 n=4+4)
WriteRTP-8                    1.52µs ± 2%     1.01µs ± 3%   -33.75%  (p=0.029 n=4+4)

name                        old speed      new speed       delta
EncryptRTP-8                71.1MB/s ± 8%  106.1MB/s ± 2%   +49.24%  (p=0.029 n=4+4)
EncryptRTPInPlace-8         72.9MB/s ± 3%  111.7MB/s ± 1%   +53.26%  (p=0.029 n=4+4)
DecryptRTP-8                19.4MB/s ± 6%   42.0MB/s ± 2%  +116.24%  (p=0.029 n=4+4)
Write-8                     68.2MB/s ± 9%  103.9MB/s ± 1%   +52.42%  (p=0.029 n=4+4)
WriteRTP-8                  73.6MB/s ± 2%  111.0MB/s ± 3%   +50.94%  (p=0.029 n=4+4)

name                        old alloc/op   new alloc/op    delta
EncryptRTP-8                    884B ± 0%       292B ± 0%   -66.97%  (p=0.029 n=4+4)
EncryptRTPInPlace-8             756B ± 0%       164B ± 0%   -78.31%  (p=0.029 n=4+4)
DecryptRTP-8                    780B ± 0%       188B ± 0%   -75.90%  (p=0.029 n=4+4)
Write-8                         916B ± 0%       324B ± 0%   -64.63%  (p=0.029 n=4+4)
WriteRTP-8                      804B ± 0%       212B ± 0%   -73.63%  (p=0.029 n=4+4)

name                        old allocs/op  new allocs/op   delta
EncryptRTP-8                    9.00 ± 0%       7.00 ± 0%   -22.22%  (p=0.029 n=4+4)
EncryptRTPInPlace-8             8.00 ± 0%       6.00 ± 0%   -25.00%  (p=0.029 n=4+4)
DecryptRTP-8                    9.00 ± 0%       7.00 ± 0%   -22.22%  (p=0.029 n=4+4)
Write-8                         9.00 ± 0%       7.00 ± 0%   -22.22%  (p=0.029 n=4+4)
WriteRTP-8                      8.00 ± 0%       6.00 ± 0%   -25.00%  (p=0.029 n=4+4)
```