@Sean-Der 2022-04-28
@jech I am all for merging it! Sorry this didn't get done sooner.

Mind adding the error checks in the tests. Just `assert.NoError()` works great for me :)