@jech 2022-04-28
Done.  (Sorry for the noise, but I don't know how to run ci-lint locally.)